backend/requirements.txt
pylint
flask
alpaca-trade-api
python-dotenv
plotly
flask-cors